+++ services/auction_service/score_ads_reactor.cc
@@ -201,6 +201,9 @@ ScoreAdsReactor::ScoreAdsReactor(
   CHECK_OK([this]() {
     PS_ASSIGN_OR_RETURN(metric_context_,
                         metric::AuctionContextMap()->Remove(request_));
+    if (log_context_.is_consented()) {
+      metric_context_->SetConsented(raw_request_.log_context().generation_id());
+    }
     return absl::OkStatus();
   }()) << "AuctionContextMap()->Get(request) should have been called";
 }