+++ services/bidding_service/bidding_service_integration_test.cc
@@ -438,7 +438,10 @@ void SetupV8Dispatcher(V8Dispatcher* dispatcher, absl::string_view adtech_js,
                        absl::string_view adtech_wasm = "") {
   ASSERT_TRUE(dispatcher->Init().ok());
   std::string wrapper_blob = GetBuyerWrappedCode(adtech_js, adtech_wasm);
-  ASSERT_TRUE(dispatcher->LoadSync("v1", wrapper_blob).ok());
+  ASSERT_TRUE(
+      dispatcher
+          ->LoadSync(kProtectedAudienceGenerateBidBlobVersion, wrapper_blob)
+          .ok());
 }
 
 absl::StatusOr<GenerateBidsRequest::GenerateBidsRawRequest>